[artf6132]: dump option screen
Adding constraints on filename and revision fields
 
* src/groovy/com/collabnet/svnedge/console/DumpBean.groovy
  (constraints): new closure. Disallows the same characters as are disallowed for repo names,
  since they are special characters for various OS or just add unnecessary complication.
  Filename is not allowed to be blank.  Revisions are restricted to positive integers.

* src/groovy/com/collabnet/svnedge/ValidationException.groovy
  New exception which can be used within a service, when validating arguments involves 
  enough internal logic that exposing it outside the service is not good for encapsulation. 

* grails-app/i18n/messages.properties
  New resources for dump parameter validation errors

* grails-app/services/com/collabnet/svnedge/console/SvnRepoService.groovy
  (createDump): throw a ValidationException, if the filename or its processing equivalent
  exists.

* grails-app/controllers/com/collabnet/svnedge/controller/RepoController.groovy
  (dumpOptions): added some return's after redirects to avoid double redirect errors.
  Don't use the automatic command object instantiation, instead if another action has
  populated and validated the DumpBean, store it through the redirect and use it here.

  (createDumpFile): Store DumpBean in flash scope, if there are validation errors.

* grails-app/views/repo/dumpOptions.gsp
  Modified validation messaging for filename and revision to be similar to the server
  admin page (new row and using two columns).  Removed other c/p error related code which 
  isn't used.


git-svn-id: https://ctf.open.collab.net/svn/repos/svnedge/trunk@2263 03e8f217-bfc6-4b7c-bcb7-0738c91e2c5f

+++ grails-app/services/com/collabnet/svnedge/console/SvnRepoService.groovy
@@ -20,6 +20,7 @@ package com.collabnet.svnedge.console
 import java.util.zip.ZipEntry;
 import java.util.zip.ZipOutputStream;
 
+import com.collabnet.svnedge.ValidationException;
 import com.collabnet.svnedge.domain.Repository 
 import com.collabnet.svnedge.domain.Server 
 import com.collabnet.svnedge.domain.ServerMode 
@@ -570,12 +571,15 @@ class SvnRepoService extends AbstractSvnEdgeService {
         File progressLogFile = 
             new File(tempLogDir, "dump-progress-" + repo.name + ".log")
         
-        log.debug("Dump command: " + cmd)
-        Process dumpProcess = commandLineService.startProcess(cmd)
-
-        FileOutputStream progress = new FileOutputStream(progressLogFile)
         File tempDumpFile = new File(dumpDir, bean.filename + "-processing")
         File finalDumpFile = new File(dumpDir, bean.filename)
+        if (tempDumpFile.exists() || finalDumpFile.exists()) {
+            throw new ValidationException("dumpBean.filename.exists", "filename")
+        }
+        
+        log.debug("Dump command: " + cmd)
+        Process dumpProcess = commandLineService.startProcess(cmd)
+        FileOutputStream progress = new FileOutputStream(progressLogFile)
         FileOutputStream out = new FileOutputStream(tempDumpFile)
 
         if (!bean.deltas && bean.filter && (bean.includePath || bean.excludePath)) {
